Put all the metrics related functions into metrics_server
